launcher.py
import sys
import os
import streamlit.web.cli as stcli

import streamlit
import pandas
import requests
import docx       
import pytesseract
import PIL        
import openpyxl   

def resolve_path(path):
    if hasattr(sys, "_MEIPASS"):
        return os.path.join(sys._MEIPASS, path)
    return os.path.join(os.getcwd(), path)

if __name__ == "__main__":
    sys.argv = [
        "streamlit",
        "run",
        resolve_path("app.py"),
        "--global.developmentMode=false",
    ]
    sys.exit(stcli.main())