requirements.txt
altair==5.5.0
altgraph==0.17.5
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Faker==37.4.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
mysql-connector-python==9.3.0
narwhals==2.15.0
numpy==2.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pillow==11.3.0
protobuf==6.33.2
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pygame==2.6.1
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
pyparsing==3.2.5
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
rdflib==7.2.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
selenium==4.36.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
xlsxwriter==3.2.9
zipp==3.23.0
